+++ HillCol.h
@@ -21,6 +21,6 @@ HillColData *hillcol_create(void);
 HillColData *hillcol_get_shared(char const *filename);
 bool hillcol_share(HillColData *hill_colours);
 DFile *hillcol_get_dfile(HillColData *hill_colours);
-size_t hillcol_get_colour(HillColData const *hill_colours, int index);
+int hillcol_get_colour(HillColData const *hill_colours, int index);
 
 #endif
