+++ OSnakes.h
@@ -30,23 +30,23 @@ typedef struct {
 
 void ObjSnakes_init(ObjSnakes *snakes_data);
 
-size_t ObjSnakes_get_count(const ObjSnakes *snakes_data);
+int ObjSnakes_get_count(const ObjSnakes *snakes_data);
 
-void ObjSnakes_get_name(const ObjSnakes *snakes_data, size_t snake,
-  char *snake_name, size_t n);
+void ObjSnakes_get_name(const ObjSnakes *snakes_data, int snake,
+  char *snake_name, int n);
 
-long int ObjSnakes_get_pal_distance(const ObjSnakes *snakes_data, size_t snake);
+long int ObjSnakes_get_pal_distance(const ObjSnakes *snakes_data, int snake);
 
-void ObjSnakes_set_pal_distance(ObjSnakes *snakes_data, size_t snake, long int distance);
+void ObjSnakes_set_pal_distance(ObjSnakes *snakes_data, int snake, long int distance);
 
 void ObjSnakes_edit(const char *tiles_set);
 
 void ObjSnakes_load(ObjSnakes *snakes_data,
-  const char *tiles_set, size_t ntiles);
+  const char *tiles_set, int ntiles);
 
 void ObjSnakes_begin_line(ObjSnakesContext *ctx,
   struct EditSession *session,
-  ObjSnakes *snakes_data, MapPoint map_pos, size_t snake,
+  ObjSnakes *snakes_data, MapPoint map_pos, int snake,
   bool inside, struct ObjEditChanges *change_info,
   struct ObjGfxMeshes *meshes);
 
@@ -56,7 +56,7 @@ void ObjSnakes_plot_line(ObjSnakesContext *ctx, MapPoint map_pos,
 void ObjSnakes_free(ObjSnakes *snakes_data);
 
 ObjRef ObjSnakes_get_value(struct EditSession *session,
-  ObjSnakes *snakes_data, MapPoint map_pos, size_t snake,
+  ObjSnakes *snakes_data, MapPoint map_pos, int snake,
   bool inside, struct ObjGfxMeshes *meshes);
 
 #endif
